+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 1.0.1 - 2021-08-22
+### Added
+- Evaluation of model on new dataset
+- New logging techniques
+- `utilities` module containing visualization techniques
+
+### Changed
+- Several image handling and data retrieval methods
+
+### Fixed
+- Reverse compactness, separability metrics
+
+
 # 1.0.0 - 2021-08-21
 ### Added
 - Initial embedding generation architecture